refactor: refactored the static assets
