+++ FrontEnd/interactive-quiz-app/app/globals.css
@@ -70,7 +70,7 @@
 }
 
 .extra-bg {
-    background-image: url("../public/stacked-waves-haikei.svg");
+    background-image: url("landingpage/stacked-waves-haikei.svg");
     background-repeat: repeat-x;
 }
