+++ FrontEnd/interactive-quiz-app/public/landingpage/laptop.jpg
