+++ FrontEnd/interactive-quiz-app/public/landingpage/single.png
