+++ FrontEnd/interactive-quiz-app/public/landingpage/triple.png
